+++ src/bun.js/bindings/ZigGlobalObject.cpp
@@ -2041,206 +2041,6 @@ JSC_DEFINE_HOST_FUNCTION(isAbortSignal, (JSGlobalObject*, CallFrame* callFrame))
     ASSERT(callFrame->argumentCount() == 1);
     return JSValue::encode(jsBoolean(callFrame->uncheckedArgument(0).inherits<JSAbortSignal>()));
 }
-static inline std::optional<JSC::JSValue> invokeReadableStreamFunction(JSC::JSGlobalObject* lexicalGlobalObject, const JSC::Identifier& identifier, JSC::JSValue thisValue, const JSC::MarkedArgumentBuffer& arguments)
-{
-    JSC::VM& vm = lexicalGlobalObject->vm();
-    auto scope = DECLARE_THROW_SCOPE(vm);
-    JSC::JSLockHolder lock(vm);
-
-    auto function = lexicalGlobalObject->get(lexicalGlobalObject, identifier);
-    scope.assertNoExceptionExceptTermination();
-    if (scope.exception()) [[unlikely]]
-        return {};
-    ASSERT(function.isCallable());
-
-    auto callData = JSC::getCallData(function);
-    auto result = JSC::call(lexicalGlobalObject, function, callData, thisValue, arguments);
-#if ASSERT_ENABLED
-    if (scope.exception()) [[unlikely]] {
-        Bun__reportError(lexicalGlobalObject, JSValue::encode(scope.exception()));
-    }
-#endif
-    EXCEPTION_ASSERT(!scope.exception() || vm.hasPendingTerminationException());
-    RETURN_IF_EXCEPTION(scope, {});
-    return result;
-}
-extern "C" bool ReadableStream__tee(JSC::EncodedJSValue possibleReadableStream, Zig::GlobalObject* globalObject, JSC::EncodedJSValue* possibleReadableStream1, JSC::EncodedJSValue* possibleReadableStream2)
-{
-    auto* readableStream = jsDynamicCast<JSReadableStream*>(JSC::JSValue::decode(possibleReadableStream));
-    if (!readableStream) [[unlikely]]
-        return false;
-
-    auto lexicalGlobalObject = globalObject;
-    auto& vm = JSC::getVM(lexicalGlobalObject);
-    auto* clientData = static_cast<JSVMClientData*>(vm.clientData);
-    auto& privateName = clientData->builtinFunctions().readableStreamInternalsBuiltins().readableStreamTeePrivateName();
-    auto scope = DECLARE_THROW_SCOPE(vm);
-
-    MarkedArgumentBuffer arguments;
-    arguments.append(readableStream);
-    arguments.append(JSC::jsBoolean(true));
-    ASSERT(!arguments.hasOverflowed());
-    auto returnedValue = invokeReadableStreamFunction(lexicalGlobalObject, privateName, JSC::jsUndefined(), arguments);
-    RETURN_IF_EXCEPTION(scope, false);
-    if (!returnedValue) return false;
-
-    auto results = convert<IDLSequence<IDLRawAny, std::array<JSValue, 2>>>(*lexicalGlobalObject, *returnedValue);
-    RETURN_IF_EXCEPTION(scope, false);
-
-    *possibleReadableStream1 = JSValue::encode(results[0]);
-    *possibleReadableStream2 = JSValue::encode(results[1]);
-    return true;
-}
-
-extern "C" void ReadableStream__cancel(JSC::EncodedJSValue possibleReadableStream, Zig::GlobalObject* globalObject)
-{
-    auto* readableStream = jsDynamicCast<JSReadableStream*>(JSC::JSValue::decode(possibleReadableStream));
-    if (!readableStream) [[unlikely]]
-        return;
-
-    if (!ReadableStream::isLocked(globalObject, readableStream)) {
-        return;
-    }
-
-    WebCore::Exception exception { AbortError };
-    ReadableStream::cancel(*globalObject, readableStream, exception);
-}
-
-extern "C" void ReadableStream__detach(JSC::EncodedJSValue possibleReadableStream, Zig::GlobalObject* globalObject)
-{
-    auto value = JSC::JSValue::decode(possibleReadableStream);
-    if (value.isEmpty() || !value.isCell())
-        return;
-
-    auto* readableStream = static_cast<JSReadableStream*>(value.asCell());
-    if (!readableStream) [[unlikely]]
-        return;
-    readableStream->setNativePtr(globalObject->vm(), jsNumber(-1));
-    readableStream->setNativeType(0);
-    readableStream->setDisturbed(true);
-}
-extern "C" bool ReadableStream__isDisturbed(JSC::EncodedJSValue possibleReadableStream, Zig::GlobalObject* globalObject);
-extern "C" bool ReadableStream__isDisturbed(JSC::EncodedJSValue possibleReadableStream, Zig::GlobalObject* globalObject)
-{
-    ASSERT(globalObject);
-    return ReadableStream::isDisturbed(globalObject, jsDynamicCast<WebCore::JSReadableStream*>(JSC::JSValue::decode(possibleReadableStream)));
-}
-
-extern "C" bool ReadableStream__isLocked(JSC::EncodedJSValue possibleReadableStream, Zig::GlobalObject* globalObject);
-extern "C" bool ReadableStream__isLocked(JSC::EncodedJSValue possibleReadableStream, Zig::GlobalObject* globalObject)
-{
-    ASSERT(globalObject);
-    WebCore::JSReadableStream* stream = jsDynamicCast<WebCore::JSReadableStream*>(JSValue::decode(possibleReadableStream));
-    return stream != nullptr && ReadableStream::isLocked(globalObject, stream);
-}
-
-extern "C" int32_t ReadableStreamTag__tagged(Zig::GlobalObject* globalObject, JSC::EncodedJSValue* possibleReadableStream, void** ptr)
-{
-    ASSERT(globalObject);
-    JSC::JSObject* object = JSValue::decode(*possibleReadableStream).getObject();
-    if (!object) {
-        *ptr = nullptr;
-        return -1;
-    }
-
-    auto& vm = JSC::getVM(globalObject);
-
-    if (!object->inherits<JSReadableStream>()) {
-        auto throwScope = DECLARE_THROW_SCOPE(vm);
-        JSValue target = object;
-        JSValue fn = JSValue();
-        auto* function = jsDynamicCast<JSC::JSFunction*>(object);
-        if (function && !function->isHostFunction() && function->jsExecutable() && function->jsExecutable()->isAsyncGenerator()) {
-            fn = object;
-            target = jsUndefined();
-        } else {
-            auto iterable = object->getIfPropertyExists(globalObject, vm.propertyNames->asyncIteratorSymbol);
-            RETURN_IF_EXCEPTION(throwScope, {});
-            if (iterable && iterable.isCallable()) {
-                fn = iterable;
-            }
-        }
-
-        if (throwScope.exception()) [[unlikely]] {
-            *ptr = nullptr;
-            return -1;
-        }
-
-        if (fn.isEmpty()) {
-            *ptr = nullptr;
-            return -1;
-        }
-
-        auto* createIterator = globalObject->builtinInternalFunctions().readableStreamInternals().m_readableStreamFromAsyncIteratorFunction.get();
-
-        JSC::MarkedArgumentBuffer arguments;
-        arguments.append(target);
-        arguments.append(fn);
-
-        JSC::JSValue result = profiledCall(globalObject, JSC::ProfilingReason::API, createIterator, JSC::getCallData(createIterator), JSC::jsUndefined(), arguments);
-
-        if (throwScope.exception()) [[unlikely]] {
-            return -1;
-        }
-
-        if (!result.isObject()) {
-            *ptr = nullptr;
-            return -1;
-        }
-
-        object = result.getObject();
-
-        ASSERT(object->inherits<JSReadableStream>());
-        *possibleReadableStream = JSValue::encode(object);
-        *ptr = nullptr;
-        ensureStillAliveHere(object);
-        return 0;
-    }
-
-    auto* readableStream = jsCast<JSReadableStream*>(object);
-
-    JSValue nativePtrHandle = readableStream->nativePtr();
-    if (nativePtrHandle.isEmpty() || !nativePtrHandle.isCell()) {
-        *ptr = nullptr;
-        return 0;
-    }
-
-    JSCell* cell = nativePtrHandle.asCell();
-
-    if (auto* casted = jsDynamicCast<JSBlobInternalReadableStreamSource*>(cell)) {
-        *ptr = casted->wrapped();
-        return 1;
-    }
-
-    if (auto* casted = jsDynamicCast<JSFileInternalReadableStreamSource*>(cell)) {
-        *ptr = casted->wrapped();
-        return 2;
-    }
-
-    if (auto* casted = jsDynamicCast<JSBytesInternalReadableStreamSource*>(cell)) {
-        *ptr = casted->wrapped();
-        return 4;
-    }
-
-    return 0;
-}
-
-extern "C" JSC::EncodedJSValue ZigGlobalObject__createNativeReadableStream(Zig::GlobalObject* globalObject, JSC::EncodedJSValue nativePtr)
-{
-    auto& vm = JSC::getVM(globalObject);
-    auto scope = DECLARE_THROW_SCOPE(vm);
-
-    auto& builtinNames = WebCore::builtinNames(vm);
-
-    auto function = globalObject->getDirect(vm, builtinNames.createNativeReadableStreamPrivateName()).getObject();
-    JSC::MarkedArgumentBuffer arguments = JSC::MarkedArgumentBuffer();
-    arguments.append(JSValue::decode(nativePtr));
-
-    auto callData = JSC::getCallData(function);
-    auto result = call(globalObject, function, callData, JSC::jsUndefined(), arguments);
-    EXCEPTION_ASSERT(!!scope.exception() == !result);
-    return JSValue::encode(result);
-}
 
 extern "C" JSC::EncodedJSValue Bun__Jest__createTestModuleObject(JSC::JSGlobalObject*);
 extern "C" JSC::EncodedJSValue Bun__Jest__testModuleObject(Zig::GlobalObject* globalObject)
@@ -2248,203 +2048,9 @@ extern "C" JSC::EncodedJSValue Bun__Jest__testModuleObject(Zig::GlobalObject* gl
     return JSValue::encode(globalObject->lazyTestModuleObject());
 }
 
-static inline JSC::EncodedJSValue ZigGlobalObject__readableStreamToArrayBufferBody(Zig::GlobalObject* globalObject, JSC::EncodedJSValue readableStreamValue)
-{
-    auto& vm = JSC::getVM(globalObject);
-
-    auto throwScope = DECLARE_THROW_SCOPE(vm);
-
-    auto* function = globalObject->m_readableStreamToArrayBuffer.get();
-    if (!function) {
-        function = JSFunction::create(vm, globalObject, static_cast<JSC::FunctionExecutable*>(readableStreamReadableStreamToArrayBufferCodeGenerator(vm)), globalObject);
-        globalObject->m_readableStreamToArrayBuffer.set(vm, globalObject, function);
-    }
-
-    JSC::MarkedArgumentBuffer arguments = JSC::MarkedArgumentBuffer();
-    arguments.append(JSValue::decode(readableStreamValue));
-
-    auto callData = JSC::getCallData(function);
-    JSValue result = call(globalObject, function, callData, JSC::jsUndefined(), arguments);
-
-    JSC::JSObject* object = result.getObject();
-
-    if (!result || result.isUndefinedOrNull()) [[unlikely]]
-        return JSValue::encode(result);
-
-    if (!object) [[unlikely]] {
-        auto throwScope = DECLARE_THROW_SCOPE(vm);
-        throwTypeError(globalObject, throwScope, "Expected object"_s);
-        return {};
-    }
-
-    JSC::JSPromise* promise = JSC::jsDynamicCast<JSC::JSPromise*>(object);
-    if (!promise) [[unlikely]] {
-        auto throwScope = DECLARE_THROW_SCOPE(vm);
-        throwTypeError(globalObject, throwScope, "Expected promise"_s);
-        return {};
-    }
-
-    RELEASE_AND_RETURN(throwScope, JSC::JSValue::encode(promise));
-}
-
-extern "C" JSC::EncodedJSValue ZigGlobalObject__readableStreamToArrayBuffer(Zig::GlobalObject* globalObject, JSC::EncodedJSValue readableStreamValue)
-{
-    return ZigGlobalObject__readableStreamToArrayBufferBody(static_cast<Zig::GlobalObject*>(globalObject), readableStreamValue);
-}
-
-extern "C" JSC::EncodedJSValue ZigGlobalObject__readableStreamToBytes(Zig::GlobalObject* globalObject, JSC::EncodedJSValue readableStreamValue)
-{
-    auto& vm = JSC::getVM(globalObject);
-
-    auto throwScope = DECLARE_THROW_SCOPE(vm);
-
-    auto* function = globalObject->m_readableStreamToBytes.get();
-    if (!function) {
-        function = JSFunction::create(vm, globalObject, static_cast<JSC::FunctionExecutable*>(readableStreamReadableStreamToBytesCodeGenerator(vm)), globalObject);
-        globalObject->m_readableStreamToBytes.set(vm, globalObject, function);
-    }
-
-    JSC::MarkedArgumentBuffer arguments = JSC::MarkedArgumentBuffer();
-    arguments.append(JSValue::decode(readableStreamValue));
-
-    auto callData = JSC::getCallData(function);
-    JSValue result = call(globalObject, function, callData, JSC::jsUndefined(), arguments);
-
-    JSC::JSObject* object = result.getObject();
-
-    if (!result || result.isUndefinedOrNull()) [[unlikely]]
-        return JSValue::encode(result);
-
-    if (!object) [[unlikely]] {
-        auto throwScope = DECLARE_THROW_SCOPE(vm);
-        throwTypeError(globalObject, throwScope, "Expected object"_s);
-        return {};
-    }
-
-    JSC::JSPromise* promise = JSC::jsDynamicCast<JSC::JSPromise*>(object);
-    if (!promise) [[unlikely]] {
-        auto throwScope = DECLARE_THROW_SCOPE(vm);
-        throwTypeError(globalObject, throwScope, "Expected promise"_s);
-        return {};
-    }
-
-    RELEASE_AND_RETURN(throwScope, JSC::JSValue::encode(promise));
-}
-
-extern "C" JSC::EncodedJSValue ZigGlobalObject__readableStreamToText(Zig::GlobalObject* globalObject, JSC::EncodedJSValue readableStreamValue)
-{
-    auto& vm = JSC::getVM(globalObject);
-
-    JSC::JSFunction* function = nullptr;
-    if (auto readableStreamToText = globalObject->m_readableStreamToText.get()) {
-        function = readableStreamToText;
-    } else {
-        function = JSFunction::create(vm, globalObject, static_cast<JSC::FunctionExecutable*>(readableStreamReadableStreamToTextCodeGenerator(vm)), globalObject);
-
-        globalObject->m_readableStreamToText.set(vm, globalObject, function);
-    }
-
-    JSC::MarkedArgumentBuffer arguments = JSC::MarkedArgumentBuffer();
-    arguments.append(JSValue::decode(readableStreamValue));
-
-    auto callData = JSC::getCallData(function);
-    return JSC::JSValue::encode(call(globalObject, function, callData, JSC::jsUndefined(), arguments));
-}
-
-extern "C" JSC::EncodedJSValue ZigGlobalObject__readableStreamToFormData(Zig::GlobalObject* globalObject, JSC::EncodedJSValue readableStreamValue, JSC::EncodedJSValue contentTypeValue)
-{
-    auto& vm = JSC::getVM(globalObject);
-
-    JSC::JSFunction* function = nullptr;
-    if (auto readableStreamToFormData = globalObject->m_readableStreamToFormData.get()) {
-        function = readableStreamToFormData;
-    } else {
-        function = JSFunction::create(vm, globalObject, static_cast<JSC::FunctionExecutable*>(readableStreamReadableStreamToFormDataCodeGenerator(vm)), globalObject);
-
-        globalObject->m_readableStreamToFormData.set(vm, globalObject, function);
-    }
-
-    JSC::MarkedArgumentBuffer arguments = JSC::MarkedArgumentBuffer();
-    arguments.append(JSValue::decode(readableStreamValue));
-    arguments.append(JSValue::decode(contentTypeValue));
-
-    auto callData = JSC::getCallData(function);
-    return JSC::JSValue::encode(call(globalObject, function, callData, JSC::jsUndefined(), arguments));
-}
-
-extern "C" JSC::EncodedJSValue ZigGlobalObject__readableStreamToJSON(Zig::GlobalObject* globalObject, JSC::EncodedJSValue readableStreamValue)
-{
-    auto& vm = JSC::getVM(globalObject);
-
-    JSC::JSFunction* function = nullptr;
-    if (auto readableStreamToJSON = globalObject->m_readableStreamToJSON.get()) {
-        function = readableStreamToJSON;
-    } else {
-        function = JSFunction::create(vm, globalObject, static_cast<JSC::FunctionExecutable*>(readableStreamReadableStreamToJSONCodeGenerator(vm)), globalObject);
-
-        globalObject->m_readableStreamToJSON.set(vm, globalObject, function);
-    }
-
-    JSC::MarkedArgumentBuffer arguments = JSC::MarkedArgumentBuffer();
-    arguments.append(JSValue::decode(readableStreamValue));
-
-    auto callData = JSC::getCallData(function);
-    return JSC::JSValue::encode(call(globalObject, function, callData, JSC::jsUndefined(), arguments));
-}
-
-extern "C" JSC::EncodedJSValue ZigGlobalObject__readableStreamToBlob(Zig::GlobalObject* globalObject, JSC::EncodedJSValue readableStreamValue)
-{
-    auto& vm = JSC::getVM(globalObject);
-
-    JSC::JSFunction* function = nullptr;
-    if (auto readableStreamToBlob = globalObject->m_readableStreamToBlob.get()) {
-        function = readableStreamToBlob;
-    } else {
-        function = JSFunction::create(vm, globalObject, static_cast<JSC::FunctionExecutable*>(readableStreamReadableStreamToBlobCodeGenerator(vm)), globalObject);
-
-        globalObject->m_readableStreamToBlob.set(vm, globalObject, function);
-    }
-
-    JSC::MarkedArgumentBuffer arguments = JSC::MarkedArgumentBuffer();
-    arguments.append(JSValue::decode(readableStreamValue));
-
-    auto callData = JSC::getCallData(function);
-    return JSC::JSValue::encode(call(globalObject, function, callData, JSC::jsUndefined(), arguments));
-}
-
 extern "C" napi_env ZigGlobalObject__makeNapiEnvForFFI(Zig::GlobalObject* globalObject)
 {
     return globalObject->makeNapiEnvForFFI();
-}
-
-JSC_DECLARE_HOST_FUNCTION(functionReadableStreamToArrayBuffer);
-JSC_DEFINE_HOST_FUNCTION(functionReadableStreamToArrayBuffer, (JSGlobalObject * globalObject, JSC::CallFrame* callFrame))
-{
-    auto& vm = JSC::getVM(globalObject);
-
-    if (callFrame->argumentCount() < 1) [[unlikely]] {
-        auto throwScope = DECLARE_THROW_SCOPE(vm);
-        throwTypeError(globalObject, throwScope, "Expected at least one argument"_s);
-        return {};
-    }
-
-    auto readableStreamValue = callFrame->uncheckedArgument(0);
-    return ZigGlobalObject__readableStreamToArrayBufferBody(static_cast<Zig::GlobalObject*>(globalObject), JSValue::encode(readableStreamValue));
-}
-
-JSC_DECLARE_HOST_FUNCTION(functionReadableStreamToBytes);
-JSC_DEFINE_HOST_FUNCTION(functionReadableStreamToBytes, (JSGlobalObject * globalObject, JSC::CallFrame* callFrame))
-{
-    auto& vm = JSC::getVM(globalObject);
-
-    if (callFrame->argumentCount() < 1) [[unlikely]] {
-        auto throwScope = DECLARE_THROW_SCOPE(vm);
-        throwTypeError(globalObject, throwScope, "Expected at least one argument"_s);
-        return {};
-    }
-
-    auto readableStreamValue = callFrame->uncheckedArgument(0);
-    return ZigGlobalObject__readableStreamToBytes(static_cast<Zig::GlobalObject*>(globalObject), JSValue::encode(readableStreamValue));
 }
 
 JSC_DEFINE_HOST_FUNCTION(jsFunctionPerformMicrotask, (JSGlobalObject * globalObject, CallFrame* callframe))
